requirements.txt
fastapi
uvicorn
openai==0.28
pinecone
requests
pyngrok
numpy
redis
aioredis
httpx
pyreadline3
python-dotenv
supabase
unidecode
pydantic-settings
pytz
tenacity